Move graph edit function implemention to a utils_graph.py to educe the size of lightray.py
